Move SPU to its own directory
