+++ src/core/iop/spu/spu_envelope.cpp
