+++ src/core/iop/spu/spu_tables.cpp
